Combining nested loop MultiLambdaParam tests.
